+++ src/main/java/com/davinci/custockspringboot/domain/repository/stock/BatchRepository.java
@@ -3,5 +3,5 @@ package com.davinci.custockspringboot.domain.repository.stock;
 import com.davinci.custockspringboot.domain.model.stock.Batch;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface BatchRepository extends JpaRepository<Batch, Long> {
+public interface BatchRepository extends JpaRepository<Batch, Integer> {
 }
